Move out model download code to separate function
